moved socket logic from server/index.js back into socket/index.js
